start moving everything into a lightcurve file
